Move platform-stm32/common into platform
