+++ platform-stm32/CMakeLists.txt
@@ -8,7 +8,6 @@ Platform convenience libraries for the STM32 microcontroller families
 
 #]=======================================================================]
 
-add_subdirectory(common)
 add_subdirectory(stm32f0xx)
 add_subdirectory(stm32f1xx)
 add_subdirectory(stm32g0xx)
