+++ platform-stm32/stm32f0xx/include/beetree/platform/stm32f0xx/serial-device-i2c.hpp
@@ -12,7 +12,7 @@
 #include <utility>
 #include "beetree/platform/i-serial-device.hpp"
 #include "beetree/platform/i-system-clock.hpp"
-#include "beetree/platform/ring-q.hpp"
+#include "beetree/platform/basics/ring-q.hpp"
 #include "beetree/platform/stm32f0xx/gpio-pin.hpp"
 #include "beetree/platform/stm32f0xx/i2c-dev-id.hpp"
